Enable (explicit engine) model checking of turn-based games (TGs).
As for LTSs, there is no explicit model type keyword added to the parser.
Instead, they are auto-detected by the absence of probabilities
combined with the presence of player information.

A new model interface, TG, is added with implementation TGSimple.
ConstructModel is extended to build TGs, and we add a new explicit
model checker: TGModelChecker, as yet uncomplete.

There is a small tweak to the parser and type checker to allow
the use of ATL properties such as <<P1>> (F "target").